refactor: Extract Class Declaration
